# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "DisqualifyIn": {
        "properties": {
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "null to reinstate",
            "title": "Reason"
          }
        },
        "title": "DisqualifyIn",
        "type": "object"
      },
...
      "ErrorOut": {
        "properties": {
          "error": {
            "description": "What went wrong, in plain words",
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
      "RankedOut": {
        "properties": {
          "adjusted": {
            "title": "Adjusted",
            "type": "number"
          },
          "interval90": {
            "items": {
              "type": "number"
            },
            "title": "Interval90",
            "type": "array"
          },
          "p_beats_next": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P Beats Next"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "raw_mean": {
            "title": "Raw Mean",
            "type": "number"
          },
          "raw_rank": {
            "title": "Raw Rank",
            "type": "integer"
          },
          "reviews": {
            "title": "Reviews",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "rank",
          "project_id",
          "title",
          "reviews",
          "raw_mean",
          "raw_rank",
          "adjusted",
          "interval90",
          "p_beats_next"
        ],
        "title": "RankedOut",
        "type": "object"
      },
      "ResultsOut": {
        "properties": {
          "excluded_reviews": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Excluded Reviews",
            "type": "array"
          },
          "judges": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Judges"
          },
          "method": {
            "additionalProperties": true,
            "title": "Method",
            "type": "object"
          },
          "ranking": {
            "items": {
              "$ref": "#/components/schemas/RankedOut"
            },
            "title": "Ranking",
            "type": "array"
          },
          "unreviewed_projects": {
            "items": {
              "type": "string"
            },
            "title": "Unreviewed Projects",
            "type": "array"
          }
        },
        "required": [
          "method",
          "ranking",
          "judges",
          "excluded_reviews",
          "unreviewed_projects"
        ],
        "title": "ResultsOut",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "An API token from /me. The session cookie works too.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "A hackathon submission and judging portal whose results you can check.",
    "title": "Plumb",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/events/import": {
      "post": {
        "description": "Body: a fixture-shaped event document (the format GET .../export returns).",
        "operationId": "import_event_api_v1_events_import_post",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not logged in: send a bearer token or a session cookie"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Logged in, but this role may not do that"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "No such thing, or not visible to you"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not now: a deadline, a lock, or a conflicting state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Invalid input"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Too many attempts; wait and retry"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Import Event",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/api/v1/events/{event_id}/results": {
      "get": {
        "description": "The normalized ranking with intervals: organizers any time, everyone after publication.",
        "operationId": "results_api_v1_events__event_id__results_get",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not logged in: send a bearer token or a session cookie"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Logged in, but this role may not do that"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "No such thing, or not visible to you"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not now: a deadline, a lock, or a conflicting state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Invalid input"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Too many attempts; wait and retry"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Results",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/disqualify": {
      "post": {
        "description": "Disqualify with a reason, or reinstate with a null reason (organizers, before publication).",
        "operationId": "disqualify_api_v1_projects__project_id__disqualify_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DisqualifyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not logged in: send a bearer token or a session cookie"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Logged in, but this role may not do that"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "No such thing, or not visible to you"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not now: a deadline, a lock, or a conflicting state"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Invalid input"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Too many attempts; wait and retry"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Disqualify",
        "tags": [
          "v1"
        ]
      }
    },
...
  }
}